requirements.txt
Django==3.1.5
djangorestframework==3.12.2
invoke==1.4.0
celery==4.4.0
flower==0.9.3
redis==3.3.11
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
raven==6.10.0
psycopg2-binary==2.8.4
django-redis==4.11.0
drf-yasg==1.20.0

requests==2.25.1
flake8==3.8.4
